# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MeshCompute Marketplace API
  version: 0.1.0
  description: >-
    Metadata-only control plane. Prompt, response, rawOutput and model content are
    intentionally absent from every central API schema.
servers:
  - url: http://127.0.0.1:4000
    description: Local API
...
paths:
...
  /v1/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      description: >-
        Deliberate divergence from doc 00 §12: without ?status the endpoint
        returns ALL providers ordered ONLINE > BUSY > OFFLINE, because the web
        marketplace exposes availability filters. Pass ?status=ONLINE for the
        spec's default behaviour.
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ProviderStatus'
      responses:
        '200':
          description: Providers ordered by availability and reputation.
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Provider'
  /v1/providers/{id}:
    get:
      tags: [Providers]
      operationId: getProvider
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Provider metadata used by the web and reliability tools.
          content:
            application/json:
              schema:
                type: object
                required: [provider]
                properties:
                  provider:
                    $ref: '#/components/schemas/Provider'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/providers/register:
    post:
      tags: [Providers]
      operationId: registerProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderRegisterRequest'
      responses:
        '201':
          description: Registered provider. The raw token is returned only here.
          content:
            application/json:
              schema:
                type: object
                required: [provider, providerToken]
                properties:
                  provider:
                    $ref: '#/components/schemas/Provider'
                  providerToken:
                    type: string
                    writeOnly: true
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/providers/{id}/heartbeat:
    post:
      tags: [Providers]
      operationId: providerHeartbeat
      security:
        - providerBearer: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Heartbeat accepted; provider is ONLINE.
          content:
            application/json:
              schema:
                type: object
                required: [provider]
                properties:
                  provider:
                    $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    providerBearer:
      type: http
      scheme: bearer
    executionBearer:
      type: http
      scheme: bearer
    executionTokenHeader:
      type: apiKey
      in: header
      name: X-Execution-Token
    sessionBearer:
      type: http
      scheme: bearer
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    ValidationError:
      description: Strict validation failed, including unknown/private fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Provider or execution token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Provider, job or route was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ProviderStatus:
      type: string
      enum: [ONLINE, OFFLINE, BUSY]
...
    ProviderRegisterRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - qvacPublicKey
        - walletAddress
        - modelKey
        - modelLabel
        - hardwareLabel
        - pricePer1kTokensAtomic
      properties:
        name:
          type: string
        qvacPublicKey:
          type: string
          minLength: 16
        walletAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
        modelKey:
          type: string
        modelLabel:
          type: string
        hardwareLabel:
          type: string
        pricePer1kTokensAtomic:
          type: string
          pattern: '^\d+$'
        pricingMode:
          const: PER_JOB
          default: PER_JOB
    Provider:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - qvacPublicKey
        - walletAddress
        - modelKey
        - modelLabel
        - hardwareLabel
        - pricePer1kTokensAtomic
        - pricingMode
        - tokenSymbol
        - status
        - reputation
        - jobsCompleted
        - jobsFailed
        - lastSeen
      properties:
        id:
          type: string
        name:
          type: string
        qvacPublicKey:
          type: string
        walletAddress:
          type: string
        modelKey:
          type: string
        modelLabel:
          type: string
        hardwareLabel:
          type: string
        description:
          type: string
          maxLength: 1000
        pricePer1kTokensAtomic:
          type: string
          pattern: '^\d+$'
        pricingMode:
          const: PER_JOB
        tokenSymbol:
          const: mUSDT
        status:
          $ref: '#/components/schemas/ProviderStatus'
        reputation:
          type: integer
          minimum: 0
          maximum: 100
        jobsCompleted:
          type: integer
          minimum: 0
        jobsFailed:
          type: integer
          minimum: 0
        lastSeen:
          type: string
          format: date-time
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...